# Inline collecting() and counting()