(*Like #39 but for `counting()` and `collecting()`*)

Measuring:

```php
use util\data\{Sequence, Aggregations};

class Operations extends \util\profiling\Measurable {
  private static $range;

  static function __static() {
    self::$range= range(0, 9999);
  }

  #[@measure]
  public function counting() {
    $count= 0;
    Sequence::of(self::$range)->counting($count)->each();
    return $count;
  }

  #[@measure]
  public function collecting() {
    $max= 0;
    Sequence::of(self::$range)->collecting($max, Aggregations::max())->each();
    return $max;
  }
}
```

Results:

```bash
# Before
$ xp -m ..\measure util.profiling.Measure Operations -n 1000
counting: 1000 iteration(s), 3.913 seconds, result= 10000
collecting: 1000 iteration(s), 7.963 seconds, result= 9999

# After
$ xp -m ..\measure util.profiling.Measure Operations -n 1000
counting: 1000 iteration(s), 0.973 seconds, result= 10000
collecting: 1000 iteration(s), 1.972 seconds, result= 9999
```

The speedup is 4.0 for both methods.